backend/requirements.txt
fastapi
uvicorn[standard]
pydub
sounddevice
wavio
SpeechRecognition
noisereduce
soundfile
transformers
textblob
sentence-transformers
reportlab
pydantic
python-dotenv
psycopg2-binary
sqlalchemy  
speechrecognition
python-multipart
openai-whisper
torch
aiofiles
databases
asyncpg
alembic
passlib[bcrypt] 
python-jose 
fastapi-users
